third_party/WebKit/LayoutTests/inspector/tracing/timeline-paint-with-style-recalc-invalidations-expected.txt

Tests the Timeline API instrumentation of paint events with style recalc invalidations.

PASS


Running: testLocalFrame

Running: testSubframe
